Modify prototypes instead of wrapping
Previously the polyfill wrapped TextEncoder and TextDecoder with
replacement implementations. This was simple but inefficient. Instead,
modify the prototypes of the original objects to reduce the size and
overhead of the polyfill.

Also update the design document to add detail about how the
implementation works, and note that the implementation has changed.

Also modify the stream-properties.html test to verify that readable and
writable accessors are enumerable. This was not the case with the
wrapping implementation, which was non-conformant with WebIDL.