Update to split API
Use constructors named TextDecoderStream and TextEncoderStream for the
streaming versions. This is based on the agreement in discussion on the
Encoding Standard, particularly
https://github.com/whatwg/encoding/issues/72#issuecomment-380101189.

The TextEncoder and TextDecoder prototypes are no longer modified, since
the new constructors are completely distinct objects. Creation of the
underlying TransformStream is no longer done lazily, as there's no
longer any concern of regressing performance for the method interfaces.

Update the tests to use the new constructors. Remove tests that involved
the encode() and decode() methods.

Update the README and design documents for the new change.